requirements.txt
flask
python-telegram-bot
Mastodon.py
requests
upstash-redis